refactor(layout): useLayoutNodes 二期拆 gitDiffPanel/panel/chrome 三段（2935→2543）
