+++ src/features/layout/hooks/useLayoutNodes.tsx
@@ -1,6 +1,4 @@
 import {
-  lazy,
-  Suspense,
   useCallback,
   useEffect,
   useLayoutEffect,
@@ -13,35 +11,23 @@ import { useEventCallback } from "../../../utils/useEventCallback";
 import { useSidebarThreadStatusProjection } from "../../threads/hooks/useSidebarThreadStatusProjection";
 import { useTranslation } from "react-i18next";
 import { buildSidebarNode } from "./layoutNodes/sidebarNode";
-import { HomeChat } from "../../home/components/HomeChat";
-import { MainHeader } from "../../app/components/MainHeader";
 import {
   CODEX_DISK_PROVIDER_PROFILE_ID,
   type CodexProviderProfileSelection,
   type CodexProviderProfileOption,
 } from "../../threads/constants/codexProviderProfiles";
-import { UpdateToast } from "../../update/components/UpdateToast";
-import { ErrorToasts } from "../../notifications/components/ErrorToasts";
 import { GlobalRuntimeNoticeDock } from "../../notifications/components/GlobalRuntimeNoticeDock";
 import type {
   ComposerNoteCardSelectionRequest,
   ComposerRewindDialogRequest,
 } from "../../composer/components/Composer";
-import { GitDiffViewer } from "../../git/components/GitDiffViewer";
 import { buildCanonicalGitChanges } from "../../git/utils/gitChangeModel";
 import {
   publishGitRepositoryActionIntent,
   type GitRepositoryActionRequest,
 } from "../../git/types/gitRepositoryActions";
 import type { GitModalPreviewRequest } from "../../git/components/GitDiffPanelTypes";
-import {
-  FileTreePanel,
-  type FileTreeRevealRequest,
-} from "../../files/components/FileTreePanel";
-import { WorkspaceFileComparePanel } from "../../files/components/WorkspaceFileComparePanel";
-import { WorkspaceSearchPanel } from "../../search/components/WorkspaceSearchPanel";
-import { PromptPanel } from "../../prompts/components/PromptPanel";
-import { ProjectMemoryPanel } from "../../project-memory/components/ProjectMemoryPanel";
+import type { FileTreeRevealRequest } from "../../files/components/FileTreePanel";
 import type {
   CanvasSemanticGraph,
   IntentCanvasCodeSelectionAnchor,
@@ -51,14 +37,10 @@ import {
   buildGitStatusProjectMapImpactInput,
   type ProjectMapImpactInput,
 } from "../../project-map/utils/impactSources";
-import { WorkspaceNoteCardPanel } from "../../note-cards/components/WorkspaceNoteCardPanel";
 import type {
   NoteCaptureDraft,
   WorkspaceNoteCaptureRequest,
 } from "../../note-cards/types";
-import { WorkspaceSessionRadarPanel } from "../../session-activity/components/WorkspaceSessionRadarPanel";
-import { TabBar } from "../../app/components/TabBar";
-import { TabletNav } from "../../app/components/TabletNav";
 import { useGlobalRuntimeNoticeDock } from "../../notifications/hooks/useGlobalRuntimeNoticeDock";
 import type { EditorNavigationLocation } from "../../app/hooks/useGitPanelController";
 import type {
@@ -92,7 +74,6 @@ import { DISABLED_WORKSPACE_SESSION_ACTIVITY } from "../../session-activity/adap
 import { useClientUiVisibility } from "../../client-ui-visibility/hooks/useClientUiVisibility";
 import {
   getHomeWorkspaceOptions,
-  resolveHomeWorkspaceId,
 } from "../../home/utils/homeWorkspaceOptions";
 import { deriveRewindWorkspaceGitState } from "./rewindWorkspaceGitState";
 import { buildWorkspaceHeaderGroups } from "./workspaceHeaderGroups";
@@ -116,6 +97,24 @@ import { useSharedSendStateRestore } from "../../shared-session/runtime/useShare
 import { buildShellRuntimeSummary } from "./layoutShellSummary";
 import { buildComposerNode } from "./layoutNodes/composerNode";
 import { buildMessagesNode } from "./layoutNodes/messagesNode";
+import { buildGitDiffPanelNode } from "./layoutNodes/gitDiffPanelNode";
+import {
+  buildGitDiffViewerNode,
+  buildFileViewPanelNode,
+  buildNoteCardsPanelNode,
+  buildFileComparePanelNode,
+  buildProjectMapPanelNode,
+  buildIntentCanvasPanelNode,
+} from "./layoutNodes/panelNodes";
+import {
+  buildHomeNode,
+  buildMainHeaderNode,
+  buildTabletNavNode,
+  buildTabBarNode,
+  buildUpdateToastNode,
+  buildErrorToastsNode,
+  buildBrowserDockNode,
+} from "./layoutNodes/chromeNodes";
 import { useCollabUiState } from "../../multi-agent/store/collabUiStore";
 import { useSharedProviderRetry } from "../../shared-session/provider-retry/useSharedProviderRetry";
 import { resolveIsSharedSession } from "../../shared-session/utils/sharedSessionIdentity";
@@ -133,35 +132,6 @@ import {
   buildTerminalDockNode,
 } from "./layoutNodeSections";
 
-const GitDiffPanel = lazy(() =>
-  import("../../git/components/GitDiffPanel").then((m) => ({
-    default: m.GitDiffPanel,
-  })),
-);
-const FileViewPanel = lazy(() =>
-  import("../../files/components/FileViewPanel").then((m) => ({
-    default: m.FileViewPanel,
-  })),
-);
-const BrowserDock = lazy(() =>
-  import("../../browser-agent/components/BrowserDock").then((m) => ({
-    default: m.BrowserDock,
-  })),
-);
-const ProjectMapPanel = lazy(() =>
-  import("../../project-map/components/ProjectMapPanel").then((m) => ({
-    default: m.ProjectMapPanel,
-  })),
-);
-const IntentCanvasManager = lazy(() =>
-  import("../../intent-canvas/components/IntentCanvasManager").then((m) => ({
-    default: m.IntentCanvasManager,
-  })),
-);
-
-function HeavyPanelFallback() {
-  return <div className="heavy-panel-fallback" aria-hidden="true" />;
-}
 
 import type {
   LayoutNodesFlatOptions,
@@ -1841,23 +1811,18 @@ export function useLayoutNodes(input: LayoutNodesOptions): LayoutNodesResult {
   const approvalToastsNode = null;
 
   const updateToastNode = useMemo(
-    () => (
-      <UpdateToast
-        state={options.updaterState}
-        onUpdate={options.onUpdate}
-        onDismiss={options.onDismissUpdate}
-      />
-    ),
+    () =>
+      buildUpdateToastNode({
+        options,
+      }),
     [options.updaterState, options.onUpdate, options.onDismissUpdate],
   );
 
   const errorToastsNode = useMemo(
-    () => (
-      <ErrorToasts
-        toasts={options.errorToasts}
-        onDismiss={options.onDismissErrorToast}
-      />
-    ),
+    () =>
+      buildErrorToastsNode({
+        options,
+      }),
     [options.errorToasts, options.onDismissErrorToast],
   );
   const homeWorkspaceOptions = useMemo(
@@ -1867,20 +1832,14 @@ export function useLayoutNodes(input: LayoutNodesOptions): LayoutNodesResult {
   );
 
   const homeNode = useMemo(
-    () => (
-      <HomeChat
-        workspaces={homeWorkspaceOptions}
-        selectedWorkspaceId={resolveHomeWorkspaceId(
-          options.activeWorkspace?.id ?? null,
-          homeWorkspaceOptions,
-        )}
-        onSelectWorkspace={options.onSelectHomeWorkspace}
-        onAddWorkspace={options.onAddWorkspace}
-        composerNode={homeComposerNode}
-        selectedEngine={homeCreationTargetEngine ?? options.selectedEngine}
-        branchControl={composerBranchControl}
-      />
-    ),
+    () =>
+      buildHomeNode({
+        options,
+        homeWorkspaceOptions,
+        homeComposerNode,
+        homeCreationTargetEngine,
+        composerBranchControl,
+      }),
     [
       homeWorkspaceOptions,
       options.activeWorkspace?.id,
@@ -1895,43 +1854,14 @@ export function useLayoutNodes(input: LayoutNodesOptions): LayoutNodesResult {
 
   const mainHeaderNode = useMemo(
     () =>
-      options.activeWorkspace ? (
-        <MainHeader
-          workspace={options.activeWorkspace}
-          parentName={options.activeParentWorkspace?.name ?? null}
-          worktreePath={
-            options.isWorktreeWorkspace ? options.activeWorkspace.path : null
-          }
-          activeFilePath={options.activeComposerFilePath}
-          openTargets={options.openAppTargets}
-          openAppIconById={options.openAppIconById}
-          selectedOpenAppId={options.selectedOpenAppId}
-          onSelectOpenAppId={options.onSelectOpenAppId}
-          sessionTabsNode={sessionTabsNode}
-          canCopyThread={canCopyActiveThread}
-          onCopyThread={options.onCopyThread}
-          onLockPanel={options.onLockPanel}
-          launchScript={options.launchScript}
-          launchScriptEditorOpen={options.launchScriptEditorOpen}
-          launchScriptDraft={options.launchScriptDraft}
-          launchScriptSaving={options.launchScriptSaving}
-          launchScriptError={options.launchScriptError}
-          onRunLaunchScript={options.onRunLaunchScript}
-          onOpenLaunchScriptEditor={options.onOpenLaunchScriptEditor}
-          onCloseLaunchScriptEditor={options.onCloseLaunchScriptEditor}
-          onLaunchScriptDraftChange={options.onLaunchScriptDraftChange}
-          onSaveLaunchScript={options.onSaveLaunchScript}
-          launchScriptsState={options.launchScriptsState}
-          showLaunchScriptControls={showTopRunControls}
-          showOpenAppMenu={showOpenWorkspaceAppControl}
-          openAppExtraActions={options.mainHeaderActions}
-          groupedWorkspaces={groupedWorkspacesForHeader}
-          activeWorkspaceId={options.activeWorkspaceId}
-          onSelectWorkspace={options.onSelectWorkspace}
-          onOpenShortcutsSettings={options.onOpenShortcutsSettings}
-
-        />
-      ) : null,
+      buildMainHeaderNode({
+        options,
+        sessionTabsNode,
+        canCopyActiveThread,
+        showTopRunControls,
+        showOpenWorkspaceAppControl,
+        groupedWorkspacesForHeader,
+      }),
     [
       options.activeWorkspace,
       options.activeParentWorkspace?.name,
@@ -1985,19 +1915,18 @@ export function useLayoutNodes(input: LayoutNodesOptions): LayoutNodesResult {
   );
 
   const tabletNavNode = useMemo(
-    () => (
-      <TabletNav
-        activeTab={options.tabletNavTab}
-        onSelect={options.onSelectTab}
-      />
-    ),
+    () =>
+      buildTabletNavNode({
+        options,
+      }),
     [options.tabletNavTab, options.onSelectTab],
   );
 
   const tabBarNode = useMemo(
-    () => (
-      <TabBar activeTab={options.activeTab} onSelect={options.onSelectTab} />
-    ),
+    () =>
+      buildTabBarNode({
+        options,
+      }),
     [options.activeTab, options.onSelectTab],
   );
   const activeWorkspaceCustomSpecRoot = useMemo(() => {
@@ -2175,401 +2104,182 @@ export function useLayoutNodes(input: LayoutNodesOptions): LayoutNodesResult {
     ],
   );
 
-  const gitDiffPanelNode = useMemo(() => {
-    if (
-      (options.filePanelMode === "files" ||
-        options.filePanelMode === "notes" ||
-        // DISABLED activity: treat residual mode as files until normalize runs
-        options.filePanelMode === "activity") &&
-      options.activeWorkspace
-    ) {
-      return (
-        <FileTreePanel
-          workspaceId={options.activeWorkspace.id}
-          workspaceName={options.activeWorkspace.name}
-          workspacePath={options.activeWorkspace.path}
-          gitRoot={options.gitRoot}
-          files={options.files}
-          directories={options.directories}
-          directoryMetadata={options.directoryMetadata}
-          sourceVersion={options.fileTreeSourceVersion}
-          isLoading={options.fileTreeLoading}
-          loadError={options.fileTreeLoadError}
-          filePanelMode="files"
-          onFilePanelModeChange={options.onFilePanelModeChange}
-          onInsertText={options.onInsertComposerText}
-          onOpenFile={options.onOpenFile}
-          onCompareFiles={options.onCompareFiles}
-          openTargets={options.openAppTargets}
-          openAppIconById={options.openAppIconById}
-          selectedOpenAppId={options.selectedOpenAppId}
-          onSelectOpenAppId={options.onSelectOpenAppId}
-          onToggleRuntimeConsole={options.onToggleRuntimeConsole}
-          isRuntimeConsoleVisible={options.runtimeConsoleVisible}
-          showSpecHubAction={false}
-          showDetachedExplorerAction={false}
-          gitStatusFiles={options.gitStatus.files}
-          gitRepositories={options.gitRepositories}
-          onGitRepositoryAction={handleFileTreeGitRepositoryAction}
-          onOpenFileHistory={options.onOpenFileHistory}
-          gitignoredFiles={options.gitignoredFiles}
-          gitignoredDirectories={options.gitignoredDirectories}
-          onRefreshFiles={options.onRefreshFiles}
-          revealRequest={fileTreeRevealRequest}
-        />
-      );
-    }
-    if (options.filePanelMode === "search") {
-      return (
-        <WorkspaceSearchPanel
-          workspaceId={options.activeWorkspace?.id ?? null}
-          filePanelMode={options.filePanelMode}
-          onFilePanelModeChange={options.onFilePanelModeChange}
-          onOpenFile={options.onOpenFile}
-        />
-      );
-    }
-    if (options.filePanelMode === "prompts") {
-      return (
-        <PromptPanel
-          prompts={options.prompts}
-          workspacePath={options.activeWorkspace?.path ?? null}
-          filePanelMode={options.filePanelMode}
-          onFilePanelModeChange={options.onFilePanelModeChange}
-          onSendPrompt={options.onSendPrompt}
-          onSendPromptToNewAgent={options.onSendPromptToNewAgent}
-          onCreatePrompt={options.onCreatePrompt}
-          onUpdatePrompt={options.onUpdatePrompt}
-          onDeletePrompt={options.onDeletePrompt}
-          onMovePrompt={options.onMovePrompt}
-          onRevealWorkspacePrompts={options.onRevealWorkspacePrompts}
-          onRevealGeneralPrompts={options.onRevealGeneralPrompts}
-          canRevealGeneralPrompts={options.canRevealGeneralPrompts}
-        />
-      );
-    }
-    if (options.filePanelMode === "memory") {
-      return (
-        <ProjectMemoryPanel
-          workspaceId={options.activeWorkspace?.id ?? null}
-          workspaces={options.workspaces}
-          onSelectWorkspace={options.onSelectWorkspace}
-          filePanelMode={options.filePanelMode}
-          onFilePanelModeChange={options.onFilePanelModeChange}
-          focusMemoryId={options.focusedProjectMemoryId ?? null}
-          focusRequestKey={options.focusedProjectMemoryRequestKey ?? 0}
-        />
-      );
-    }
-    if (options.filePanelMode === "radar") {
-      return (
-        <WorkspaceSessionRadarPanel
-          runningSessions={options.sessionRadarRunningSessions}
-          recentCompletedSessions={options.sessionRadarRecentCompletedSessions}
-          onSelectThread={options.onSelectThread}
-        />
-      );
-    }
-    return (
-      <Suspense fallback={<HeavyPanelFallback />}>
-        <GitDiffPanel
-          workspaceId={options.activeWorkspace?.id ?? null}
-          workspacePath={options.activeWorkspace?.path ?? null}
-          headerControlsTarget={gitModeControlsTarget}
-          mode={options.gitPanelMode}
-          onModeChange={options.onGitPanelModeChange}
-          onOpenGitHistoryPanel={options.onOpenGitHistoryPanel}
-          isGitHistoryOpen={options.appMode === "gitHistory"}
-          diffEntries={options.gitDiffs}
-          gitDiffListView={options.gitDiffListView}
-          onGitDiffListViewChange={options.onGitDiffListViewChange}
-          toggleGitDiffListViewShortcut={options.toggleGitDiffListViewShortcut}
-          filePanelMode={options.filePanelMode}
-          onFilePanelModeChange={options.onFilePanelModeChange}
-          worktreeApplyLabel={options.worktreeApplyLabel}
-          worktreeApplyTitle={options.worktreeApplyTitle}
-          worktreeApplyLoading={options.worktreeApplyLoading}
-          worktreeApplyError={options.worktreeApplyError}
-          worktreeApplySuccess={options.worktreeApplySuccess}
-          onApplyWorktreeChanges={options.onApplyWorktreeChanges}
-          branchName={
-            options.gitStatus.branchName || t("workspace.unknownBranch")
-          }
-          totalAdditions={canonicalGitPanelTotals.additions}
-          totalDeletions={canonicalGitPanelTotals.deletions}
-          fileStatus={options.fileStatus}
-          diffViewStyle={options.gitDiffViewStyle}
-          onDiffViewStyleChange={options.onGitDiffViewStyleChange}
-          error={options.gitStatus.error}
-          logError={options.gitLogError}
-          logLoading={options.gitLogLoading}
-          stagedFiles={canonicalGitPanelChanges.stagedFiles}
-          unstagedFiles={canonicalGitPanelChanges.unstagedFiles}
-          onSelectFile={options.onSelectDiff}
-          onOpenFile={(path, repositoryRoot) =>
-            options.onOpenFile(path, undefined, {
-              pathDomain: "git",
-              repositoryRoot,
-            })
-          }
-          onOpenFileHistory={options.onOpenFileHistory}
-          modalPreviewRequest={gitModalPreviewRequest}
-          selectedPath={sidebarSelectedDiffPath}
-          logEntries={options.gitLogEntries}
-          logTotal={options.gitLogTotal}
-          logAhead={options.gitLogAhead}
-          logBehind={options.gitLogBehind}
-          logAheadEntries={options.gitLogAheadEntries}
-          logBehindEntries={options.gitLogBehindEntries}
-          logUpstream={options.gitLogUpstream}
-          selectedCommitSha={options.selectedCommitSha}
-          onSelectCommit={options.onSelectCommit}
-          issues={options.gitIssues}
-          issuesTotal={options.gitIssuesTotal}
-          issuesLoading={options.gitIssuesLoading}
-          issuesError={options.gitIssuesError}
-          pullRequests={options.gitPullRequests}
-          pullRequestsTotal={options.gitPullRequestsTotal}
-          pullRequestsLoading={options.gitPullRequestsLoading}
-          pullRequestsError={options.gitPullRequestsError}
-          selectedPullRequest={options.selectedPullRequestNumber}
-          onSelectPullRequest={options.onSelectPullRequest}
-          gitRemoteUrl={options.gitRemoteUrl}
-          gitRoot={options.gitRoot}
-          gitRootCandidates={options.gitRootCandidates}
-          gitRootScanDepth={options.gitRootScanDepth}
-          gitRootScanLoading={options.gitRootScanLoading}
-          gitRootScanError={options.gitRootScanError}
-          gitRootScanHasScanned={options.gitRootScanHasScanned}
-          onGitRootScanDepthChange={options.onGitRootScanDepthChange}
-          onScanGitRoots={options.onScanGitRoots}
-          onSelectGitRoot={options.onSelectGitRoot}
-          onClearGitRoot={options.onClearGitRoot}
-          onPickGitRoot={options.onPickGitRoot}
-          onStageAllChanges={options.onStageGitAll}
-          onStageFile={options.onStageGitFile}
-          onUnstageAllChanges={options.onUnstageGitAll}
-          onUnstageFile={options.onUnstageGitFile}
-          onUnstageFiles={options.onUnstageGitPaths}
-          onRevertFile={options.onRevertGitFile}
-          onRevertFiles={options.onRevertGitPaths}
-          onRevertAllChanges={options.onRevertAllGitChanges}
-          commitMessage={options.commitMessage}
-          commitMessageLoading={options.commitMessageLoading}
-          commitMessageError={options.commitMessageError}
-          onCommitMessageChange={options.onCommitMessageChange}
-          onGenerateCommitMessage={options.onGenerateCommitMessage}
-          onCommit={options.onCommit}
-          onCommitAndPush={options.onCommitAndPush}
-          onCommitAndSync={options.onCommitAndSync}
-          onPush={options.onPush}
-          onSync={options.onSync}
-          commitLoading={options.commitLoading}
-          pushLoading={options.pushLoading}
-          syncLoading={options.syncLoading}
-          commitError={options.commitError}
-          pushError={options.pushError}
-          syncError={options.syncError}
-          commitsAhead={options.commitsAhead}
-          multiRepositoryMode={options.multiRepositoryMode}
-          repositoryStatuses={options.repositoryStatuses}
-          repositoryStatusesLoading={options.repositoryStatusesLoading}
-          onRefreshRepositoryStatuses={options.onRefreshRepositoryStatuses}
-          onStageRepositoryFile={options.onStageRepositoryFile}
-          onUnstageRepositoryFile={options.onUnstageRepositoryFile}
-          onUnstageRepositoryAll={options.onUnstageRepositoryAll}
-          onUnstageRepositoryFiles={options.onUnstageRepositoryFiles}
-          onRevertRepositoryFile={options.onRevertRepositoryFile}
-          onRevertRepositoryFiles={options.onRevertRepositoryFiles}
-          onStageRepositoryAll={options.onStageRepositoryAll}
-          onCommitRepositories={options.onCommitRepositories}
-          repositoryCommitSummary={options.repositoryCommitSummary}
-          onRefreshGitStatus={options.queueGitStatusRefresh}
-          onRefreshGitLog={options.refreshGitLog}
-          onRefreshGitDiffs={options.refreshGitDiffs}
-          onCreateCodeAnnotation={handleCreateCodeAnnotation}
-          onRemoveCodeAnnotation={handleRemoveCodeAnnotation}
-          codeAnnotations={selectedCodeAnnotations}
-        />
-      </Suspense>
-    );
-  }, [
-    options.filePanelMode,
-    options.activeWorkspace,
-    options.gitRoot,
-    options.files,
-    options.directories,
-    options.directoryMetadata,
-    options.fileTreeSourceVersion,
-    options.fileTreeLoading,
-    options.fileTreeLoadError,
-    options.onFilePanelModeChange,
-    options.onInsertComposerText,
-    options.onOpenFile,
-    options.onCompareFiles,
-    options.openAppTargets,
-    options.openAppIconById,
-    options.selectedOpenAppId,
-    options.onSelectOpenAppId,
-    options.onToggleRuntimeConsole,
-    options.runtimeConsoleVisible,
-    options.gitStatus.files,
-    options.gitStatus.branchName,
-    options.gitStatus.error,
-    options.gitRepositories,
-    handleFileTreeGitRepositoryAction,
-    options.onOpenFileHistory,
-    options.gitignoredFiles,
-    options.gitignoredDirectories,
-    options.onRefreshFiles,
-    fileTreeRevealRequest,
-    options.prompts,
-    options.onSendPrompt,
-    options.onSendPromptToNewAgent,
-    options.onCreatePrompt,
-    options.onUpdatePrompt,
-    options.onDeletePrompt,
-    options.onMovePrompt,
-    options.onRevealWorkspacePrompts,
-    options.onRevealGeneralPrompts,
-    options.canRevealGeneralPrompts,
-    options.workspaces,
-    options.onSelectWorkspace,
-    options.focusedProjectMemoryId,
-    options.focusedProjectMemoryRequestKey,
-    options.sessionRadarRunningSessions,
-    options.sessionRadarRecentCompletedSessions,
-    options.onSelectThread,
-    gitModeControlsTarget,
-    options.gitPanelMode,
-    options.onGitPanelModeChange,
-    options.onOpenGitHistoryPanel,
-    options.appMode,
-    options.gitDiffs,
-    options.gitDiffListView,
-    options.onGitDiffListViewChange,
-    options.toggleGitDiffListViewShortcut,
-    options.worktreeApplyLabel,
-    options.worktreeApplyTitle,
-    options.worktreeApplyLoading,
-    options.worktreeApplyError,
-    options.worktreeApplySuccess,
-    options.onApplyWorktreeChanges,
-    t,
-    canonicalGitPanelTotals.additions,
-    canonicalGitPanelTotals.deletions,
-    options.fileStatus,
-    options.gitDiffViewStyle,
-    options.onGitDiffViewStyleChange,
-    options.gitLogError,
-    options.gitLogLoading,
-    canonicalGitPanelChanges.stagedFiles,
-    canonicalGitPanelChanges.unstagedFiles,
-    options.onSelectDiff,
-    gitModalPreviewRequest,
-    sidebarSelectedDiffPath,
-    options.gitLogEntries,
-    options.gitLogTotal,
-    options.gitLogAhead,
-    options.gitLogBehind,
-    options.gitLogAheadEntries,
-    options.gitLogBehindEntries,
-    options.gitLogUpstream,
-    options.selectedCommitSha,
-    options.onSelectCommit,
-    options.gitIssues,
-    options.gitIssuesTotal,
-    options.gitIssuesLoading,
-    options.gitIssuesError,
-    options.gitPullRequests,
-    options.gitPullRequestsTotal,
-    options.gitPullRequestsLoading,
-    options.gitPullRequestsError,
-    options.selectedPullRequestNumber,
-    options.onSelectPullRequest,
-    options.gitRemoteUrl,
-    options.gitRootCandidates,
-    options.gitRootScanDepth,
-    options.gitRootScanLoading,
-    options.gitRootScanError,
-    options.gitRootScanHasScanned,
-    options.onGitRootScanDepthChange,
-    options.onScanGitRoots,
-    options.onSelectGitRoot,
-    options.onClearGitRoot,
-    options.onPickGitRoot,
-    options.onStageGitAll,
-    options.onStageGitFile,
-    options.onUnstageGitAll,
-    options.onUnstageGitFile,
-    options.onUnstageGitPaths,
-    options.onRevertGitFile,
-    options.onRevertGitPaths,
-    options.onRevertAllGitChanges,
-    options.commitMessage,
-    options.commitMessageLoading,
-    options.commitMessageError,
-    options.onCommitMessageChange,
-    options.onGenerateCommitMessage,
-    options.onCommit,
-    options.onCommitAndPush,
-    options.onCommitAndSync,
-    options.onPush,
-    options.onSync,
-    options.commitLoading,
-    options.pushLoading,
-    options.syncLoading,
-    options.commitError,
-    options.pushError,
-    options.syncError,
-    options.commitsAhead,
-    options.multiRepositoryMode,
-    options.repositoryStatuses,
-    options.repositoryStatusesLoading,
-    options.onRefreshRepositoryStatuses,
-    options.onStageRepositoryFile,
-    options.onUnstageRepositoryFile,
-    options.onUnstageRepositoryAll,
-    options.onUnstageRepositoryFiles,
-    options.onRevertRepositoryFile,
-    options.onRevertRepositoryFiles,
-    options.onStageRepositoryAll,
-    options.onCommitRepositories,
-    options.repositoryCommitSummary,
-    options.queueGitStatusRefresh,
-    options.refreshGitLog,
-    options.refreshGitDiffs,
-    handleCreateCodeAnnotation,
-    handleRemoveCodeAnnotation,
-    selectedCodeAnnotations,
-  ]);
+  const gitDiffPanelNode = useMemo(
+    () =>
+      buildGitDiffPanelNode({
+        options,
+        handleFileTreeGitRepositoryAction,
+        fileTreeRevealRequest,
+        gitModeControlsTarget,
+        t,
+        canonicalGitPanelTotals,
+        canonicalGitPanelChanges,
+        gitModalPreviewRequest,
+        sidebarSelectedDiffPath,
+        handleCreateCodeAnnotation,
+        handleRemoveCodeAnnotation,
+        selectedCodeAnnotations,
+      }),
+    [
+      options.filePanelMode,
+      options.activeWorkspace,
+      options.gitRoot,
+      options.files,
+      options.directories,
+      options.directoryMetadata,
+      options.fileTreeSourceVersion,
+      options.fileTreeLoading,
+      options.fileTreeLoadError,
+      options.onFilePanelModeChange,
+      options.onInsertComposerText,
+      options.onOpenFile,
+      options.onCompareFiles,
+      options.openAppTargets,
+      options.openAppIconById,
+      options.selectedOpenAppId,
+      options.onSelectOpenAppId,
+      options.onToggleRuntimeConsole,
+      options.runtimeConsoleVisible,
+      options.gitStatus.files,
+      options.gitStatus.branchName,
+      options.gitStatus.error,
+      options.gitRepositories,
+      handleFileTreeGitRepositoryAction,
+      options.onOpenFileHistory,
+      options.gitignoredFiles,
+      options.gitignoredDirectories,
+      options.onRefreshFiles,
+      fileTreeRevealRequest,
+      options.prompts,
+      options.onSendPrompt,
+      options.onSendPromptToNewAgent,
+      options.onCreatePrompt,
+      options.onUpdatePrompt,
+      options.onDeletePrompt,
+      options.onMovePrompt,
+      options.onRevealWorkspacePrompts,
+      options.onRevealGeneralPrompts,
+      options.canRevealGeneralPrompts,
+      options.workspaces,
+      options.onSelectWorkspace,
+      options.focusedProjectMemoryId,
+      options.focusedProjectMemoryRequestKey,
+      options.sessionRadarRunningSessions,
+      options.sessionRadarRecentCompletedSessions,
+      options.onSelectThread,
+      gitModeControlsTarget,
+      options.gitPanelMode,
+      options.onGitPanelModeChange,
+      options.onOpenGitHistoryPanel,
+      options.appMode,
+      options.gitDiffs,
+      options.gitDiffListView,
+      options.onGitDiffListViewChange,
+      options.toggleGitDiffListViewShortcut,
+      options.worktreeApplyLabel,
+      options.worktreeApplyTitle,
+      options.worktreeApplyLoading,
+      options.worktreeApplyError,
+      options.worktreeApplySuccess,
+      options.onApplyWorktreeChanges,
+      t,
+      canonicalGitPanelTotals.additions,
+      canonicalGitPanelTotals.deletions,
+      options.fileStatus,
+      options.gitDiffViewStyle,
+      options.onGitDiffViewStyleChange,
+      options.gitLogError,
+      options.gitLogLoading,
+      canonicalGitPanelChanges.stagedFiles,
+      canonicalGitPanelChanges.unstagedFiles,
+      options.onSelectDiff,
+      gitModalPreviewRequest,
+      sidebarSelectedDiffPath,
+      options.gitLogEntries,
+      options.gitLogTotal,
+      options.gitLogAhead,
+      options.gitLogBehind,
+      options.gitLogAheadEntries,
+      options.gitLogBehindEntries,
+      options.gitLogUpstream,
+      options.selectedCommitSha,
+      options.onSelectCommit,
+      options.gitIssues,
+      options.gitIssuesTotal,
+      options.gitIssuesLoading,
+      options.gitIssuesError,
+      options.gitPullRequests,
+      options.gitPullRequestsTotal,
+      options.gitPullRequestsLoading,
+      options.gitPullRequestsError,
+      options.selectedPullRequestNumber,
+      options.onSelectPullRequest,
+      options.gitRemoteUrl,
+      options.gitRootCandidates,
+      options.gitRootScanDepth,
+      options.gitRootScanLoading,
+      options.gitRootScanError,
+      options.gitRootScanHasScanned,
+      options.onGitRootScanDepthChange,
+      options.onScanGitRoots,
+      options.onSelectGitRoot,
+      options.onClearGitRoot,
+      options.onPickGitRoot,
+      options.onStageGitAll,
+      options.onStageGitFile,
+      options.onUnstageGitAll,
+      options.onUnstageGitFile,
+      options.onUnstageGitPaths,
+      options.onRevertGitFile,
+      options.onRevertGitPaths,
+      options.onRevertAllGitChanges,
+      options.commitMessage,
+      options.commitMessageLoading,
+      options.commitMessageError,
+      options.onCommitMessageChange,
+      options.onGenerateCommitMessage,
+      options.onCommit,
+      options.onCommitAndPush,
+      options.onCommitAndSync,
+      options.onPush,
+      options.onSync,
+      options.commitLoading,
+      options.pushLoading,
+      options.syncLoading,
+      options.commitError,
+      options.pushError,
+      options.syncError,
+      options.commitsAhead,
+      options.multiRepositoryMode,
+      options.repositoryStatuses,
+      options.repositoryStatusesLoading,
+      options.onRefreshRepositoryStatuses,
+      options.onStageRepositoryFile,
+      options.onUnstageRepositoryFile,
+      options.onUnstageRepositoryAll,
+      options.onUnstageRepositoryFiles,
+      options.onRevertRepositoryFile,
+      options.onRevertRepositoryFiles,
+      options.onStageRepositoryAll,
+      options.onCommitRepositories,
+      options.repositoryCommitSummary,
+      options.queueGitStatusRefresh,
+      options.refreshGitLog,
+      options.refreshGitDiffs,
+      handleCreateCodeAnnotation,
+      handleRemoveCodeAnnotation,
+      selectedCodeAnnotations,
+    ],
+  );
 
   const gitDiffViewerNode = useMemo(
-    () => (
-      <GitDiffViewer
-        workspaceId={options.activeWorkspace?.id ?? null}
-        diffs={options.gitDiffs}
-        listView={options.gitDiffListView}
-        selectedPath={options.selectedDiffPath}
-        scrollRequestId={options.diffScrollRequestId}
-        isLoading={options.gitDiffLoading}
-        error={options.gitDiffError}
-        diffStyle={options.gitDiffViewStyle}
-        alignedTextPreview
-        onDiffStyleChange={options.onGitDiffViewStyleChange}
-        pullRequest={options.selectedPullRequest}
-        pullRequestComments={options.selectedPullRequestComments}
-        pullRequestCommentsLoading={options.selectedPullRequestCommentsLoading}
-        pullRequestCommentsError={options.selectedPullRequestCommentsError}
-        onActivePathChange={options.onDiffActivePathChange}
-        onOpenFile={options.onOpenFile}
-        onRequestClose={options.onExitDiff}
-        onCreateCodeAnnotation={handleCreateCodeAnnotation}
-        onRemoveCodeAnnotation={handleRemoveCodeAnnotation}
-        codeAnnotations={selectedCodeAnnotations}
-        codeAnnotationSurface="embedded-diff-view"
-      />
-    ),
+    () =>
+      buildGitDiffViewerNode({
+        options,
+        handleCreateCodeAnnotation,
+        handleRemoveCodeAnnotation,
+        selectedCodeAnnotations,
+      }),
     [
       options.activeWorkspace?.id,
       options.gitDiffs,
@@ -2597,73 +2307,17 @@ export function useLayoutNodes(input: LayoutNodesOptions): LayoutNodesResult {
   // props Object.is 比较，让根 render 反复重渲染已挂载的重面板（S1 残余项）。
   const fileViewPanelNode = useMemo(
     () =>
-      options.editorFilePath && options.activeWorkspace ? (
-      <Suspense fallback={<HeavyPanelFallback />}>
-        <FileViewPanel
-          workspaceId={options.activeWorkspace.id}
-          workspaceName={options.activeWorkspace.name}
-          workspacePath={options.activeWorkspace.path}
-          gitRoot={options.gitRoot}
-          gitRepositories={options.gitRepositories}
-          customSpecRoot={activeWorkspaceCustomSpecRoot}
-          filePath={options.editorFilePath}
-          navigationTarget={options.editorNavigationTarget}
-          highlightMarkers={
-            options.editorHighlightTarget?.path === options.editorFilePath
-              ? options.editorHighlightTarget.markers
-              : null
-          }
-          gitStatusFiles={options.gitStatus.files}
-          openTabs={options.openEditorTabs}
-          activeTabPath={options.editorFilePath}
-          onActivateTab={options.onActivateEditorTab}
-          onCloseTab={options.onCloseEditorTab}
-          onCloseOtherTabs={options.onCloseOtherEditorTabs}
-          onCloseAllTabs={options.onCloseAllEditorTabs}
-          onReorderTabs={options.onReorderEditorTabs}
-          fileReferenceMode={options.fileReferenceMode}
-          onFileReferenceModeChange={options.onFileReferenceModeChange}
-          activeFileLineRange={options.activeComposerFileLineRange}
-          onActiveFileLineRangeChange={options.onActiveEditorLineRangeChange}
-          onActiveCodeAnchorChange={options.onActiveCodeSelectionAnchorChange}
-          onAssociateIntentCanvasCodeAnchor={
-            handleAssociateIntentCanvasCodeAnchor
-          }
-          openTargets={options.openAppTargets}
-          openAppIconById={options.openAppIconById}
-          selectedOpenAppId={options.selectedOpenAppId}
-          onSelectOpenAppId={options.onSelectOpenAppId}
-          editorSplitLayout={options.editorSplitLayout}
-          onToggleEditorSplitLayout={options.onToggleEditorSplitLayout}
-          isEditorFileMaximized={options.isEditorFileMaximized}
-          onToggleEditorFileMaximized={options.onToggleEditorFileMaximized}
-          onNavigateToLocation={options.onOpenFile}
-          onOpenFileHistory={options.onOpenFileHistory}
-          onRevealInFileTree={handleRevealInFileTree}
-          onClose={options.onExitEditor}
-          onInsertText={options.onInsertComposerText}
-          onCreateCodeAnnotation={handleCreateCodeAnnotation}
-          onCaptureNote={handleCaptureWorkspaceNote}
-          onRemoveCodeAnnotation={handleRemoveCodeAnnotation}
-          codeAnnotations={selectedCodeAnnotations}
-          externalChangeMonitoringEnabled={
-            options.externalChangeMonitoringEnabled
-          }
-          externalChangeTransportMode={options.externalChangeTransportMode}
-          externalChangeApplyMode={options.externalChangeApplyMode}
-          externalChangeAutoApplyDebounceMs={
-            options.externalChangeAutoApplyDebounceMs
-          }
-          markdownPreviewSnapshotMode={
-            options.liveEditPreviewEnabled ? "live" : "stable"
-          }
-          fileRenderPressure={fileRenderPressure}
-          saveFileShortcut={options.saveFileShortcut}
-          findInFileShortcut={options.findInFileShortcut}
-          expandSelectionShortcut={options.expandSelectionShortcut}
-        />
-      </Suspense>
-      ) : null,
+      buildFileViewPanelNode({
+        options,
+        activeWorkspaceCustomSpecRoot,
+        handleAssociateIntentCanvasCodeAnchor,
+        handleRevealInFileTree,
+        handleCreateCodeAnnotation,
+        handleCaptureWorkspaceNote,
+        handleRemoveCodeAnnotation,
+        selectedCodeAnnotations,
+        fileRenderPressure,
+      }),
     [
       options.editorFilePath,
       options.activeWorkspace,
@@ -2720,19 +2374,14 @@ export function useLayoutNodes(input: LayoutNodesOptions): LayoutNodesResult {
       options.editorSplitCompanion === "notes");
   const noteCardsPanelNode = useMemo(
     () =>
-      isWorkspaceNoteCardsMounted ? (
-        <WorkspaceNoteCardPanel
-          workspaceId={options.activeWorkspace?.id ?? null}
-          workspaceName={options.activeWorkspace?.name ?? null}
-          workspacePath={options.activeWorkspace?.path ?? null}
-          focusNoteId={options.focusedWorkspaceNoteId ?? null}
-          focusRequestKey={options.focusedWorkspaceNoteRequestKey ?? 0}
-          captureRequest={workspaceNoteCaptureRequest}
-          onCaptureRequestHandled={handleWorkspaceNoteCaptureRequestHandled}
-          onReferenceNote={handleReferenceWorkspaceNote}
-          onOpenCodeSource={handleOpenWorkspaceNoteCodeSource}
-        />
-      ) : null,
+      buildNoteCardsPanelNode({
+        options,
+        isWorkspaceNoteCardsMounted,
+        workspaceNoteCaptureRequest,
+        handleWorkspaceNoteCaptureRequestHandled,
+        handleReferenceWorkspaceNote,
+        handleOpenWorkspaceNoteCodeSource,
+      }),
     [
       isWorkspaceNoteCardsMounted,
       options.activeWorkspace,
@@ -2747,16 +2396,9 @@ export function useLayoutNodes(input: LayoutNodesOptions): LayoutNodesResult {
 
   const fileComparePanelNode = useMemo(
     () =>
-      options.centerMode === "fileCompare" ? (
-        <WorkspaceFileComparePanel
-          session={options.fileCompareSession}
-          workspaceId={options.activeWorkspace?.id ?? null}
-          workspaceName={options.activeWorkspace?.name ?? null}
-          workspacePath={options.activeWorkspace?.path ?? null}
-          saveFileShortcut={options.saveFileShortcut}
-          onClose={options.onCloseFileCompare}
-        />
-      ) : null,
+      buildFileComparePanelNode({
+        options,
+      }),
     [
       options.centerMode,
       options.fileCompareSession,
@@ -2775,25 +2417,12 @@ export function useLayoutNodes(input: LayoutNodesOptions): LayoutNodesResult {
   );
   const projectMapPanelNode = useMemo(
     () =>
-      isProjectMapSurfaceActive ? (
-        <Suspense fallback={<HeavyPanelFallback />}>
-          <ProjectMapPanel
-            key={options.activeWorkspace?.id ?? "no-workspace"}
-            activeWorkspace={options.activeWorkspace ?? null}
-            workspaceName={options.activeWorkspace?.name ?? null}
-            selectedEngine={options.selectedEngine ?? null}
-            selectedModelId={options.selectedModelId}
-            models={options.models}
-            datasetController={options.projectMapDatasetController}
-            changedFilePaths={projectMapImpactInput.filePaths}
-            changedFileSource={projectMapImpactInput.source}
-            activeCodeSelectionAnchor={options.activeCodeSelectionAnchor}
-            onOpenEvidenceFile={handleOpenProjectMapEvidenceFile}
-            onOpenIntentCanvas={options.onOpenIntentCanvas}
-            onOpenIntentCanvasFromRelationship={options.onOpenIntentCanvas}
-          />
-        </Suspense>
-      ) : null,
+      buildProjectMapPanelNode({
+        options,
+        isProjectMapSurfaceActive,
+        projectMapImpactInput,
+        handleOpenProjectMapEvidenceFile,
+      }),
     [
       isProjectMapSurfaceActive,
       options.activeWorkspace,
@@ -2810,19 +2439,11 @@ export function useLayoutNodes(input: LayoutNodesOptions): LayoutNodesResult {
 
   const intentCanvasPanelNode = useMemo(
     () =>
-      isIntentCanvasSurfaceActive ? (
-        <Suspense fallback={<HeavyPanelFallback />}>
-          <IntentCanvasManager
-            activeWorkspace={options.activeWorkspace ?? null}
-            activeThreadId={options.activeThreadId ?? null}
-            openRequest={options.intentCanvasOpenRequest ?? null}
-            onOpenRequestConsumed={options.onIntentCanvasOpenRequestConsumed}
-            onAttachToThread={options.onAttachIntentCanvasToThread}
-            onOpenProjectMap={options.onOpenProjectMap}
-            onOpenSourceFile={handleOpenProjectMapEvidenceFile}
-          />
-        </Suspense>
-      ) : null,
+      buildIntentCanvasPanelNode({
+        options,
+        isIntentCanvasSurfaceActive,
+        handleOpenProjectMapEvidenceFile,
+      }),
     [
       isIntentCanvasSurfaceActive,
       options.activeWorkspace,
@@ -2885,20 +2506,7 @@ export function useLayoutNodes(input: LayoutNodesOptions): LayoutNodesResult {
     diffLabel: t("workspace.diff"),
     onBackFromDiff: options.onBackFromDiff,
   });
-  const browserDockNode =
-    options.browserDockOpen &&
-    options.centerMode === "chat" &&
-    options.activeWorkspaceId ? (
-      <Suspense fallback={<HeavyPanelFallback />}>
-        <BrowserDock
-          workspaceId={options.activeWorkspaceId}
-          ownerSurface="main-split-browser-dock"
-          displayMode="embedded"
-          className="browser-agent-center-panel-dock"
-          onClosePanel={options.onCloseBrowserDock}
-        />
-      </Suspense>
-    ) : null;
+  const browserDockNode = buildBrowserDockNode({ options });
 
   return {
     codeAnnotationBridgeProps,
